Move tests from Clob to text classes. Move db-specific conditions.
Took tests from ClobTypeTest and moved them to either VarcharTypeTest (new class) or NVarcharTypeTest (new class).
Moved database-specific logic from the ClobType class and moved it to VarcharType or NVarcharType.
Some Clob groovy tests need to be moved to other groovy classes
